% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar.R
\name{bar}
\alias{bar}
\title{bar function}
\usage{
bar(ts)
}
\arguments{
\item{ts}{Argument to specify number of datapoints. Numeric or string value
allowed.}
}
\description{
Tells a bar joke.
}
\details{
This demo function tells a bar joke.
}
\examples{
    bar(ts=3)
}
\author{
Yves R. Sagaert
}
