% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diva_to_dits.R
\name{diva_to_dits}
\alias{diva_to_dits}
\title{diva_to_dits function}
\usage{
diva_to_dits(
  file = "",
  data = "",
  sep = ";",
  start = c(1, 1),
  frequency = 1,
  d = "yearweek",
  i = "sku",
  v = "real_sales"
)
}
\arguments{
\item{file}{Filename of csv file to read. Alternatively, a datatable is also
taken.}

\item{data}{Datatable to convert directly, without reading a .csv-file.}

\item{sep}{Separator for the csv file: often , or ;}

\item{start}{Start of time series in time series format: for monthly data
c(2025,12), for weekly data c(2025, 40), for daily data c(2025,11,25).}

\item{frequency}{Frequency of the time series}

\item{d}{Column name for date information}

\item{i}{Column name for item information}

\item{v}{Column name for value information}
}
\description{
Format from DIVA to DITS.
}
\details{
The function converts from DIVA (Date Item Value Attributes) to DITS (Date
Item Table Structure, or Date Item TimeSeries) format.
}
\examples{
 \dontrun{
    diva_to_dits("data.csv")
 }
}
\author{
Yves R. Sagaert
}
