% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsdummy_package.R
\docType{package}
\name{tsdummy}
\alias{tsdummy-package}
\alias{tsdummy}
\title{tsdummy: Time Series Dummy}
\description{
This package creates extra features for data with date information of in a
time series format. Binary dummies or summarising feature engineering.
}
\author{
Yves R. Sagaert \email{yves.r.sagaert@gmail.com}
}
