% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_info.R
\docType{package}
\name{tsdummy}
\alias{tsuntangle}
\alias{tsuntangle-package}
\alias{tsdummy}
\title{tsuntangle: Time Series Untangle Package}
\description{
This package creates clear formatting, extra features,
dummies for data with a date information or in time series format.
}
\author{
Yves R. Sagaert \email{yves.r.sagaert@gmail.com}
}
